MLStyle is devonly
