+++ TyPython/src/CPython.NumPy.DevOnly.compiled.jl
@@ -1,0 +1,13 @@
+# This file is automatically generated by DevOnly.jl
+import Serialization
+import SHA
+let
+    expected_sign = b"\x52\x7e\x49\xc2\xe2\x6e\xca\xf5\x84\xd5\xb7\x47\x59\xe0\x19\xae\xed\x74\xce\x00\xad\x32\x1b\x5a\x12\x35\x2f\x90"
+    data = b"\x37\x4a\x4c\x0f\x04\x00\x00\x00\x16\x02\x01\x08\x74\x6f\x70\x6c\x65\x76\x65\x6c\x17\xe0\x01\x3d\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x43\x50\x79\x74\x68\x6f\x6e\x2e\x4e\x75\x6d\x50\x79\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x01\x08\x66\x75\x6e\x63\x74\x69\x6f\x6e\x16\x02\x56\x01\x0c\x66\x72\x6f\x6d\x5f\x6e\x64\x61\x72\x72\x61\x79\x16\x02\x57\x7d\x01\x02\x50\x79\x16\x1b\x01\x05\x62\x6c\x6f\x63\x6b\x17\xe8\x2c\x02\x00\x17\xe9\x2c\x02\x00\x16\x02\x59\x01\x02\x6e\x70\x16\x01\x56\x01\x09\x67\x65\x74\x5f\x6e\x75\x6d\x70\x79\x17\xea\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x03\x56\x01\x02\x21\x3d\x16\x03\x56\x16\x02\x01\x01\x2e\x01\x05\x50\x79\x41\x50\x49\x1a\x01\x10\x50\x79\x4f\x62\x6a\x65\x63\x74\x5f\x48\x61\x73\x41\x74\x74\x72\x7d\x16\x02\x56\x01\x1c\x61\x74\x74\x72\x69\x62\x75\x74\x65\x5f\x73\x79\x6d\x62\x6f\x6c\x5f\x74\x6f\x5f\x70\x79\x6f\x62\x6a\x65\x63\x74\x1a\x01\x10\x5f\x5f\x61\x72\x72\x61\x79\x5f\x73\x74\x72\x75\x63\x74\x5f\x5f\xe0\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\xeb\x2c\x02\x00\x16\x02\x56\x01\x05\x65\x72\x72\x6f\x72\x16\x03\x01\x06\x73\x74\x72\x69\x6e\x67\x39\x21\x0e\x66\x72\x6f\x6d\x5f\x6e\x64\x61\x72\x72\x61\x79\x3a\x20\x7d\x39\x21\x18\x20\x68\x61\x73\x20\x6e\x6f\x20\x5f\x5f\x61\x72\x72\x61\x79\x5f\x73\x74\x72\x75\x63\x74\x5f\x5f\x17\xed\x2c\x02\x00\x16\x02\x59\x2c\x14\x00\x16\x02\x01\x01\x2e\x7d\x1a\x2c\x14\x00\x17\xee\x2c\x02\x00\x16\x02\x59\x01\x03\x70\x74\x72\x16\x02\x56\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x16\x02\x01\x01\x2e\x5f\x1a\x54\x01\x10\x50\x79\x41\x72\x72\x61\x79\x49\x6e\x74\x65\x72\x66\x61\x63\x65\x16\x02\x56\x01\x10\x5f\x67\x65\x74\x5f\x63\x61\x70\x73\x75\x6c\x65\x5f\x70\x74\x72\x2c\x14\x00\x17\xef\x2c\x02\x00\x16\x02\x59\x01\x04\x69\x6e\x66\x6f\x16\x02\x57\x16\x01\x01\x03\x72\x65\x66\x01\x03\x70\x74\x72\x2c\x20\x00\x17\xf0\x2c\x02\x00\x16\x02\x01\x02\x7c\x7c\x16\x03\x56\x01\x02\x69\x6e\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x08\x74\x79\x70\x65\x6b\x69\x6e\x64\x01\x0f\x73\x75\x70\x70\x6f\x72\x74\x65\x64\x5f\x6b\x69\x6e\x64\x73\x16\x02\x56\x01\x05\x65\x72\x72\x6f\x72\x16\x02\x01\x06\x73\x74\x72\x69\x6e\x67\x39\x21\x19\x75\x6e\x73\x75\x70\x70\x6f\x72\x74\x65\x64\x20\x6e\x75\x6d\x70\x79\x20\x64\x74\x79\x70\x65\x3a\x20\x16\x02\x56\x01\x04\x43\x68\x61\x72\x16\x02\x01\x01\x2e\x01\x03\x70\x74\x72\x1a\x2c\x29\x00\x17\xf1\x2c\x02\x00\x16\x02\x59\x01\x05\x66\x6c\x61\x67\x73\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x05\x66\x6c\x61\x67\x73\x17\xf2\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x7c\x7c\x16\x02\x56\x9c\x16\x02\x01\x02\x7c\x7c\x16\x03\x56\x01\x08\x63\x68\x65\x63\x6b\x62\x69\x74\x01\x05\x66\x6c\x61\x67\x73\x01\x16\x4e\x50\x59\x5f\x41\x52\x52\x41\x59\x5f\x46\x5f\x43\x4f\x4e\x54\x49\x47\x55\x4f\x55\x53\x16\x03\x56\x2c\x37\x00\x01\x05\x66\x6c\x61\x67\x73\x01\x16\x4e\x50\x59\x5f\x41\x52\x52\x41\x59\x5f\x43\x5f\x43\x4f\x4e\x54\x49\x47\x55\x4f\x55\x53\x16\x02\x56\x9c\x16\x03\x56\x2c\x37\x00\x01\x05\x66\x6c\x61\x67\x73\x01\x1e\x54\x59\x50\x59\x5f\x53\x55\x50\x50\x4f\x52\x54\x45\x44\x5f\x4e\x4f\x5f\x43\x4f\x50\x59\x5f\x4e\x50\x5f\x46\x4c\x41\x47\x16\x0a\x01\x05\x62\x6c\x6f\x63\x6b\x17\xf4\x2c\x02\x00\x16\x02\x59\x7d\x16\x03\x56\x16\x02\x01\x01\x2e\x01\x02\x6e\x70\x1a\x01\x04\x63\x6f\x70\x79\x7d\x16\x02\x01\x02\x6b\x77\x01\x05\x6f\x72\x64\x65\x72\x16\x03\x56\x01\x07\x70\x79\x5f\x63\x61\x73\x74\x01\x02\x50\x79\x21\x01\x46\x17\xf5\x2c\x02\x00\x16\x02\x59\x2c\x14\x00\x16\x02\x01\x01\x2e\x7d\x1a\x2c\x14\x00\x17\xf6\x2c\x02\x00\x16\x02\x59\x01\x03\x70\x74\x72\x16\x02\x56\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x16\x02\x01\x01\x2e\x5f\x1a\x54\x2c\x20\x00\x16\x02\x56\x2c\x22\x00\x2c\x14\x00\x17\xf7\x2c\x02\x00\x16\x02\x59\x01\x04\x69\x6e\x66\x6f\x16\x02\x57\x16\x01\x01\x03\x72\x65\x66\x01\x03\x70\x74\x72\x2c\x20\x00\x17\xf8\x2c\x02\x00\x16\x02\x59\x01\x05\x66\x6c\x61\x67\x73\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x05\x66\x6c\x61\x67\x73\x17\xfc\x2c\x02\x00\x16\x02\x59\x01\x05\x73\x68\x61\x70\x65\x16\x02\x57\x16\x02\x56\x53\x16\x02\x01\x09\x67\x65\x6e\x65\x72\x61\x74\x6f\x72\x16\x02\x56\x01\x03\x49\x6e\x74\x6e\x16\x02\x59\x6e\x16\x05\x56\x01\x0b\x75\x6e\x73\x61\x66\x65\x5f\x77\x72\x61\x70\x16\x01\x01\x0a\x70\x61\x72\x61\x6d\x65\x74\x65\x72\x73\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x0b\x50\x79\x5f\x69\x6e\x74\x70\x74\x72\x5f\x74\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x05\x73\x68\x61\x70\x65\x16\x02\x56\x01\x03\x49\x6e\x74\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x02\x6e\x64\x01\x09\x53\x68\x61\x70\x65\x54\x79\x70\x65\x17\xff\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x2c\x37\x00\x01\x05\x66\x6c\x61\x67\x73\x2c\x3a\x00\x16\x02\x56\x9c\x16\x03\x56\x2c\x37\x00\x01\x05\x66\x6c\x61\x67\x73\x2c\x38\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x21\x00\x00\x00\x2c\x02\x00\x16\x02\x59\x01\x05\x73\x68\x61\x70\x65\x16\x02\x56\x01\x07\x72\x65\x76\x65\x72\x73\x65\x01\x05\x73\x68\x61\x70\x65\x17\x31\x24\x00\x00\x00\x2c\x02\x00\x16\x02\x59\x01\x03\x61\x72\x72\x16\x08\x01\x05\x62\x6c\x6f\x63\x6b\x4d\x16\x02\x59\x01\x05\x23\x23\x37\x33\x36\x16\x02\xa4\x16\x02\x56\x01\x04\x43\x68\x61\x72\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x2c\x29\x00\x16\x02\x56\x01\x03\x49\x6e\x74\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x08\x69\x74\x65\x6d\x73\x69\x7a\x65\x16\x02\x01\x02\x69\x66\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x33\x36\x10\x53\x1f\x4e\x9b\x44\x02\x00\x00\x00\x00\x0f\x00\x08\x16\x1e\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x25\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x62\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe0\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x0c\x23\x23\x72\x65\x74\x75\x72\x6e\x23\x37\x33\x34\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x26\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x01\x0d\x72\x65\x67\x69\x73\x74\x65\x72\x5f\x72\x6f\x6f\x74\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x04\x42\x6f\x6f\x6c\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x01\x0c\x73\x79\x6d\x62\x6f\x6c\x69\x63\x67\x6f\x74\x6f\x01\x11\x23\x23\x23\x23\x66\x69\x6e\x61\x6c\x23\x37\x33\x35\x23\x37\x33\x39\x17\x31\x28\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x69\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe0\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x29\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x04\x49\x6e\x74\x38\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x2b\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x69\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe1\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x2c\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x05\x49\x6e\x74\x31\x36\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x2e\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x69\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe3\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x2f\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x05\x49\x6e\x74\x33\x32\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x31\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x69\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe7\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x32\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x05\x49\x6e\x74\x36\x34\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x34\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x66\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe1\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x35\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x07\x46\x6c\x6f\x61\x74\x31\x36\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x37\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x66\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe3\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x38\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x07\x46\x6c\x6f\x61\x74\x33\x32\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x3a\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x66\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe7\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x3b\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x07\x46\x6c\x6f\x61\x74\x36\x34\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x3d\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x75\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe0\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x3e\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x05\x55\x49\x6e\x74\x38\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x40\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x75\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe1\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x41\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x06\x55\x49\x6e\x74\x31\x36\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x43\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x75\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe3\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x44\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x06\x55\x49\x6e\x74\x33\x32\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x46\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x75\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe7\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x47\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x06\x55\x49\x6e\x74\x36\x34\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x49\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x63\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe3\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x4a\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x0a\x43\x6f\x6d\x70\x6c\x65\x78\x46\x31\x36\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x4c\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x63\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xe7\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x4d\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x0a\x43\x6f\x6d\x70\x6c\x65\x78\x46\x33\x32\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x4f\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x0f\x63\x16\x03\x56\x5b\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\xef\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x50\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x2c\x83\x00\x7d\x16\x05\x56\x2c\x58\x00\x16\x01\x2c\x5a\x00\x16\x02\x01\x02\x6b\x77\x01\x03\x6f\x77\x6e\x4c\x50\x16\x03\x56\xa0\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x54\x01\x0a\x43\x6f\x6d\x70\x6c\x65\x78\x46\x36\x34\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x04\x64\x61\x74\x61\x01\x05\x73\x68\x61\x70\x65\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x17\x31\x52\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x01\x03\x69\x73\x61\x01\x05\x23\x23\x37\x33\x36\x12\x00\x1c\x01\x04\x76\x61\x72\x32\x47\x46\x12\x00\x1c\x01\x04\x76\x61\x72\x31\x47\x46\x10\x53\x1f\x4e\x9b\x44\x02\x00\x00\x00\x2c\x82\x01\x2c\x81\x01\x16\x03\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x05\x23\x23\x37\x33\x37\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe0\x16\x02\x59\x01\x05\x23\x23\x37\x33\x38\x16\x02\x01\x03\x72\x65\x66\x01\x05\x23\x23\x37\x33\x36\xe1\x4d\x16\x04\x01\x05\x62\x6c\x6f\x63\x6b\x16\x02\x59\x01\x06\x6e\x62\x79\x74\x65\x73\x01\x05\x23\x23\x37\x33\x38\x16\x02\x59\x01\x04\x63\x6f\x64\x65\x01\x05\x23\x23\x37\x33\x37\x16\x02\x59\x2c\x80\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x53\x00\x00\x00\x2c\x02\x00\x16\x02\x56\x01\x05\x74\x68\x72\x6f\x77\x16\x02\x56\x01\x05\x65\x72\x72\x6f\x72\x16\x04\x01\x06\x73\x74\x72\x69\x6e\x67\x2c\x2d\x00\x01\x04\x63\x6f\x64\x65\x21\x01\x20\x01\x06\x6e\x62\x79\x74\x65\x73\x16\x01\x2c\x8b\x00\x2c\x8c\x00\x16\x02\x56\x34\x10\x01\x06\x23\x65\x72\x72\x6f\x72\x1f\x4e\x9e\x44\x39\x21\x62\x6d\x61\x74\x63\x68\x69\x6e\x67\x20\x6e\x6f\x6e\x2d\x65\x78\x68\x61\x75\x73\x74\x69\x76\x65\x2c\x20\x61\x74\x20\x23\x3d\x20\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x43\x50\x79\x74\x68\x6f\x6e\x2e\x4e\x75\x6d\x50\x79\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x3a\x33\x36\x20\x3d\x23\x16\x01\x01\x0d\x73\x79\x6d\x62\x6f\x6c\x69\x63\x6c\x61\x62\x65\x6c\x2c\x8c\x00\x2c\x80\x00\x17\x31\x56\x00\x00\x00\x2c\x02\x00\x16\x02\x01\x02\x69\x66\x16\x02\x01\x02\x26\x26\x16\x03\x56\x2c\x37\x00\x01\x05\x66\x6c\x61\x67\x73\x2c\x3a\x00\x16\x02\x56\x9c\x16\x03\x56\x2c\x37\x00\x01\x05\x66\x6c\x61\x67\x73\x2c\x38\x00\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x57\x00\x00\x00\x2c\x02\x00\x16\x03\x01\x02\x69\x66\x16\x03\x56\x5a\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x02\x6e\x64\xe1\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x58\x00\x00\x00\x2c\x02\x00\x16\x02\x59\x01\x03\x61\x72\x72\x16\x02\x56\x01\x09\x74\x72\x61\x6e\x73\x70\x6f\x73\x65\x01\x03\x61\x72\x72\x16\x02\x01\x06\x65\x6c\x73\x65\x69\x66\x16\x02\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x59\x00\x00\x00\x2c\x02\x00\x16\x03\x56\x01\x01\x3e\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x02\x6e\x64\xe1\x16\x04\x01\x05\x62\x6c\x6f\x63\x6b\x17\x31\x5a\x00\x00\x00\x2c\x02\x00\x16\x02\x59\x01\x04\x70\x65\x72\x6d\x16\x02\x56\x53\x16\x02\x56\x01\x07\x72\x65\x76\x65\x72\x73\x65\x16\x03\x56\x01\x01\x3a\xe0\x16\x02\x01\x01\x2e\x01\x04\x69\x6e\x66\x6f\x1a\x01\x02\x6e\x64\x17\x31\x5b\x00\x00\x00\x2c\x02\x00\x16\x02\x59\x01\x03\x61\x72\x72\x16\x03\x56\x01\x11\x50\x65\x72\x6d\x75\x74\x65\x64\x44\x69\x6d\x73\x41\x72\x72\x61\x79\x01\x03\x61\x72\x72\x01\x04\x70\x65\x72\x6d\x17\x31\x5e\x00\x00\x00\x2c\x02\x00\x16\x01\x55\x01\x03\x61\x72\x72\x16\x05\x56\x34\x10\x01\x05\x23\x64\x6f\x63\x21\x1f\x4e\x9e\x01\x04\x44\x6f\x63\x73\x44\x1f\x0b\x1b\x88\xd0\xb7\x27\x9d\x5e\x9a\x69\x4c\x7d\x23\xa2\x66\x45\x9c\x01\x08\x54\x79\x50\x79\x74\x68\x6f\x6e\x01\x07\x43\x50\x79\x74\x68\x6f\x6e\x44\x16\x03\x56\x10\x01\x07\x42\x69\x6e\x64\x69\x6e\x67\x1f\x4e\x9e\x01\x04\x44\x6f\x63\x73\x44\x1f\x0b\x1b\x88\xd0\xb7\x27\x9d\x5e\x9a\x69\x4c\x7d\x23\xa2\x66\x45\x9c\x2c\xb4\x01\x01\x07\x43\x50\x79\x74\x68\x6f\x6e\x44\x1a\x2c\x07\x00\x16\x03\x56\x34\x10\x01\x07\x23\x64\x6f\x63\x73\x74\x72\x1f\x4e\x9e\x01\x04\x44\x6f\x63\x73\x44\x16\x02\x56\x34\x10\x01\x05\x23\x73\x76\x65\x63\x1f\x4e\x9b\x44\x39\x30\x5c\x01\x00\x00\x00\x00\x00\x00\x20\x20\x20\x20\x66\x72\x6f\x6d\x5f\x6e\x64\x61\x72\x72\x61\x79\x28\x78\x3a\x3a\x50\x79\x29\x0a\x0a\x43\x6f\x6e\x76\x65\x72\x74\x20\x6e\x75\x6d\x70\x79\x20\x6e\x64\x61\x72\x72\x61\x79\x20\x74\x6f\x20\x6a\x75\x6c\x69\x61\x20\x61\x72\x72\x61\x79\x2c\x0a\x6e\x64\x61\x72\x72\x61\x79\x20\x77\x69\x74\x68\x20\x46\x2d\x63\x6f\x6e\x74\x69\x67\x75\x6f\x75\x73\x20\x63\x6f\x75\x6c\x64\x20\x62\x65\x20\x63\x6f\x6e\x76\x65\x72\x74\x65\x64\x20\x74\x6f\x20\x41\x72\x72\x61\x79\x20\x77\x69\x74\x68\x6f\x75\x74\x20\x63\x6f\x70\x79\x2c\x0a\x6e\x64\x61\x72\x72\x61\x79\x20\x77\x69\x74\x68\x20\x43\x2d\x63\x6f\x6e\x74\x69\x67\x75\x6f\x75\x73\x20\x63\x6f\x75\x6c\x64\x20\x62\x65\x20\x63\x6f\x6e\x76\x65\x72\x74\x65\x64\x20\x74\x6f\x20\x54\x72\x61\x6e\x73\x70\x6f\x73\x65\x20\x6f\x66\x20\x41\x72\x72\x61\x79\x28\x32\x44\x29\x20\x6f\x72\x20\x50\x65\x72\x6d\x75\x74\x65\x64\x44\x69\x6d\x73\x41\x72\x72\x61\x79\x20\x6f\x66\x20\x41\x72\x72\x61\x79\x28\x68\x69\x67\x68\x20\x64\x69\x6d\x65\x6e\x73\x69\x6f\x6e\x73\x29\x20\x77\x69\x74\x68\x6f\x75\x74\x20\x63\x6f\x70\x79\x2c\x0a\x6f\x74\x68\x65\x72\x20\x6e\x64\x61\x72\x72\x61\x79\x20\x77\x69\x6c\x6c\x20\x66\x69\x72\x73\x74\x20\x62\x65\x20\x63\x6f\x70\x69\x65\x64\x20\x74\x6f\x20\x66\x2d\x63\x6f\x6e\x74\x69\x67\x75\x6f\x75\x73\x20\x6e\x64\x61\x72\x72\x61\x79\x2c\x20\x74\x68\x65\x6e\x20\x63\x6f\x6e\x76\x65\x72\x74\x65\x64\x20\x74\x6f\x20\x41\x72\x72\x61\x79\x2e\x0a\x16\x04\x56\x10\x01\x04\x44\x69\x63\x74\x1f\x4e\x9e\x44\x02\x00\x00\x00\x00\x01\x46\x34\x10\x01\x04\x50\x61\x69\x72\x1f\x4e\x9b\x44\x02\x00\x00\x00\x00\x01\x00\x21\x01\x04\x70\x61\x74\x68\x39\x21\x3d\x45\x3a\x5c\x50\x72\x6f\x6a\x65\x63\x74\x73\x5c\x47\x69\x74\x34\x5c\x6a\x6e\x75\x6d\x70\x79\x5c\x54\x79\x50\x79\x74\x68\x6f\x6e\x5c\x73\x72\x63\x5c\x43\x50\x79\x74\x68\x6f\x6e\x2e\x4e\x75\x6d\x50\x79\x2e\x44\x65\x76\x4f\x6e\x6c\x79\x2e\x6a\x6c\x34\x10\x01\x04\x50\x61\x69\x72\x1f\x4e\x9b\x44\x02\x00\x00\x00\x00\x01\x00\x08\x01\x0a\x6c\x69\x6e\x65\x6e\x75\x6d\x62\x65\x72\xe0\x34\x10\x01\x04\x50\x61\x69\x72\x1f\x4e\x9b\x44\x02\x00\x00\x00\x00\x01\x00\x1f\x01\x06\x6d\x6f\x64\x75\x6c\x65\x1f\x0b\x1b\x88\xd0\xb7\x27\x9d\x5e\x9a\x69\x4c\x7d\x23\xa2\x66\x45\x9c\x2c\xb4\x01\x01\x07\x43\x50\x79\x74\x68\x6f\x6e\x44\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x01\x05\x55\x6e\x69\x6f\x6e\x16\x02\x01\x05\x63\x75\x72\x6c\x79\x53\x01\x02\x50\x79"
+    sign_data = SHA.sha224(data)
+    if sign_data != expected_sign
+        error("The compiled data file", "E:\\Projects\\Git4\\jnumpy\\TyPython\\src\\CPython.NumPy.DevOnly.compiled.jl", "is not verified. Please recompile it.")
+    end
+    buf = IOBuffer(data)
+    Base.eval(@__MODULE__(), Serialization.deserialize(buf))
+end
