+++ TyPython/src/TyPython.jl
@@ -1,6 +1,7 @@
 module TyPython
 
 export CPython
+include("DevOnly.jl")
 include("Reflection.jl")
 include("Utils.jl")
 include("C.jl")
